Fix build on macOS
Avoid a duplicate definition of GetKernelVersion. Move CheckKernelVersion
to still be available on all Unixes.
